Initial code commit.
This code is not any good. It is just a quick refactor of some existing code.

At the moment, it still relies on the caller knowing a lot about the structure
of the report, like internal IDs and the exact filter string and parameters XML.

But it works.

This commit servers as the starting point for future improvements.